refactor: update User & Dealer inherit
