+++ review_v1_updates.md
@@ -33,8 +33,9 @@
 - [x] **14. 일급 컬렉션을 사용하도록 수정**
 - [x] **15. 모든 엔티티를 작게 유지할 것.**
   - 일급 컬렉션이나 원시 값 포장 등을 사용.
-- [ ] **16. 딜러와 플레이어에서 발생하는 중복 코드 제거하기**
+- [x] **16. 딜러와 플레이어에서 발생하는 중복 코드 제거하기**
   - 먼저 공부할 것. https://tecoble.techcourse.co.kr/post/2020-05-18-inheritance-vs-composition/
+  - participant와 user는 is-a 이기 때문에 상속으로 수정.
 - [x] **17. 블랙잭 도메인에 대한 테스트 코드를 진행할 것**
 - [x] **18. 테스트 커버리지 80%를 목표로**
 - [x] **19. Parser와 Validator의 구분 명확하게 할 것.**
