+++ src/main/java/domain/Dealer.java
@@ -2,31 +2,13 @@ package domain;
 
 import vo.GameResult;
 
-public class Dealer {
+public class Dealer extends Participant{
     private final static Integer WINNING_SCORE_BOUNDARY = 21;
-
-    private final Hand hand;
-
-    public Dealer() {
-        this.hand = new Hand();
-    }
-
-    public void receiveCard(Card card) {
-        hand.saveCard(card);
-    }
-
-    public String getCardsDisplay() {
-        return hand.getCardsDisplay();
-    }
 
     public Boolean determineDealerDealMore() {
         return hand.determineDealerDealMore();
     }
-
-    public String getDealerFinalDisplay() {
-        return hand.getFinalDisplay();
-    }
-
+    
     public GameResult judgeUserResult(int userTotalScore) {
         if (hand.calculateTotalScore() == WINNING_SCORE_BOUNDARY) {
             return GameResult.WIN;
